feat: Moved graphql app creation into fastapi as it is fastapi specific.
