+++ graphql/src/adapters/fastapi/graphql_app.py
@@ -1,6 +1,5 @@
 # noinspection PyPackageRequirements
-from strawberry.fastapi import GraphQLRouter  # todo: pull out like https://git.earthdata.nasa.gov/projects/ORCA/repos/orca-prototype-graphql/browse/src/server/adapters/api/graphql/adapter_interfaces/fastapi.py?at=refs%2Fheads%2Ffeature%2Ffastapi-strawberry-clean
-
+from strawberry.fastapi import GraphQLRouter
 from src.adapters.graphql import graphql_settings
 from src.adapters.graphql.schemas import get_schema
